dev-requirements.txt
Django>=1.6.0
django-atomic-signals>=1.0.0,<2.0
django-celery-with-redis>=3.0,<4.0
psycopg2
flake8
django-nose
rednose
